refactor : 반환값 VO 인터페이스로 통일 (#797)
