+++ src/main/java/com/clubber/ClubberServer/domain/review/controller/KeywordController.java
@@ -3,14 +3,15 @@ package com.clubber.ClubberServer.domain.review.controller;
 import com.clubber.ClubberServer.domain.review.service.ReviewService;
 import com.clubber.ClubberServer.global.config.swagger.DisableSwaggerSecurity;
 import com.clubber.ClubberServer.global.infrastructure.outer.perspective.client.PerspectiveClient;
-import com.clubber.ClubberServer.global.vo.enums.EnumMapperVO;
+import com.clubber.ClubberServer.global.mapper.enums.EnumMapperVO;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+
+import java.util.List;
 
 @RestController
 @RequestMapping("/api/v1/keywords")
@@ -28,5 +29,4 @@ public class KeywordController {
     public List<EnumMapperVO> getTotalReviews() {
         return reviewService.getTotalKeywords();
     }
-
 }
