+++ src/main/java/com/clubber/ClubberServer/domain/review/domain/Keyword.java
@@ -1,7 +1,6 @@
 package com.clubber.ClubberServer.domain.review.domain;
 
 import com.clubber.ClubberServer.global.mapper.enums.EnumDefaultMapperType;
-import com.clubber.ClubberServer.global.vo.enums.EnumDefaultMapperVO;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import java.util.Arrays;
 import lombok.AllArgsConstructor;
